+++ lib/mesh_gradient.dart
@@ -2,5 +2,9 @@ library mesh_gradient;
 
 export 'package:mesh_gradient/src/animated_mesh_gradient.dart';
 export 'package:mesh_gradient/src/animated_mesh_gradient_options.dart';
-export 'package:mesh_gradient/src/mesh_gradient.dart';
-export 'package:mesh_gradient/src/mesh_gradient_point.dart';
+
+//dev
+
+// export 'package:mesh_gradient/src/mesh_gradient.dart';
+// export 'package:mesh_gradient/src/mesh_gradient_point.dart';
+// export 'package:mesh_gradient/src/mesh_gradient_options.dart';