+++ lib/src/mesh_gradient.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_shaders/flutter_shaders.dart';
+import 'package:mesh_gradient/src/mesh_gradient_options.dart';
 import 'package:mesh_gradient/src/mesh_gradient_painter.dart';
 import 'package:mesh_gradient/src/mesh_gradient_point.dart';
 
@@ -7,11 +8,13 @@ class MeshGradient extends StatefulWidget {
   const MeshGradient({
     super.key,
     required this.points,
+    required this.options,
     this.child,
   });
 
   final List<MeshGradientPoint> points;
   final Widget? child;
+  final MeshGradientOptions options;
 
   @override
   State<MeshGradient> createState() => _MeshGradientState();
@@ -32,6 +35,7 @@ class _MeshGradientState extends State<MeshGradient> {
           painter: MeshGradientPainter(
             shader: shader,
             points: widget.points,
+            options: widget.options,
           ),
           child: child,
         );