+++ lib/src/mesh_gradient_options.dart
@@ -1,0 +1,7 @@
+class MeshGradientOptions {
+  MeshGradientOptions({
+    this.blend = 3,
+  });
+
+  final double blend;
+}